src/requirements.txt
# Core ML & Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Fairness & Bias Detection
aif360>=0.5.0
fairlearn>=0.9.0

# Causal Inference
networkx>=3.1
causal-learn>=0.1.3

# LLM Integration (DeepSeek via OpenAI-compatible API)
requests>=2.31.0
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
pyvis>=0.3.2

# Web UI
streamlit>=1.28.0

# Utilities
pyyaml>=6.0
